api: rename apikite -> apiutil
